compilation ok, without textures

+++ srcs/maths/tler.c
@@ -6,7 +6,7 @@
 /*   By: ajeannin <marvin@42.fr>                    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/03/01 15:32:16 by ajeannin          #+#    #+#             */
-/*   Updated: 2024/03/12 17:01:01 by ajeannin         ###   ########.fr       */
+/*   Updated: 2024/03/13 19:38:38 by ajeannin         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -18,11 +18,11 @@
 //
 //have to find beta from alpha (angle of the ray), so we could use * and not /
 //(use less ressources)
-int get_ipx(t_player *p, int i_py, double angle)
+int get_ipx(t_player *p, double i_py, double angle)
 {
 	int	i_px;
 	int i_px2;
-	int i_px3;
+	double i_px3;
 	double res;
 
 	i_px = 0;
@@ -72,11 +72,11 @@ int get_ipx(t_player *p, int i_py, double angle)
 }
 
 //same for coll
-int get_ipy(t_player *p, int i_px, double angle)
+int get_ipy(t_player *p, double i_px, double angle)
 {
 	int	i_py;
 	int i_py2;
-	int i_py3;
+	double i_py3;
 	double res;
 
 	i_py = 0;
